+++ src/SIL.XForge.Scripture/ClientApp/src/app/app.component.spec.ts
@@ -27,7 +27,6 @@ import { LocationService } from 'xforge-common/location.service';
 import { UserDoc } from 'xforge-common/models/user-doc';
 import { NoticeService } from 'xforge-common/notice.service';
 import { OnlineStatusService } from 'xforge-common/online-status.service';
-import { PermissionsService } from 'xforge-common/permissions.service';
 import { PwaService } from 'xforge-common/pwa.service';
 import { QueryParameters } from 'xforge-common/query-parameters';
 import { TestOnlineStatusModule } from 'xforge-common/test-online-status.module';
@@ -44,6 +43,7 @@ import { CheckingQuestionsService } from './checking/checking/checking-questions
 import { QuestionDoc } from './core/models/question-doc';
 import { SFProjectProfileDoc } from './core/models/sf-project-profile-doc';
 import { SF_TYPE_REGISTRY } from './core/models/sf-type-registry';
+import { PermissionsService } from './core/permissions.service';
 import { SFProjectService } from './core/sf-project.service';
 import { NavigationProjectSelectorComponent } from './navigation-project-selector/navigation-project-selector.component';
 import { NmtDraftAuthGuard, SettingsAuthGuard, SyncAuthGuard, UsersAuthGuard } from './shared/project-router.guard';
